Implement BufferedPacketQueue
This class is used to buffer packets which might arrive to VideoPlayer
out of order.
The exact reason of the failed ordering is not known at the moment.
One of the strongest hypotheses is that UDP messes up the packet order,
but we don't know exactly which UDP socket: one of multiple sockets on
the Air Unit, or one of multiple sockets on the Ground Station

But this workaround helps to mitigate an extremely annoying decoding
problem.